requirements.txt
boto3>=1.34.0
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
networkx>=3.1
matplotlib>=3.7.0
plotly>=5.17.0
streamlit-drawable-canvas>=0.9.0
